backend/requirements.txt
flask
torch
torchvision
Pillow
numpy
opencv-python
pandas